Add alloc_page function
